refactor: 테스트의 EntityCreateUtil 위치 변경
